Crate corrosiff

Corrosiff

corrosiff is a library for reading data from .siff files or converting the intensity data in .siff files to .tiff files if you don’t care about arrival times and just want images that can be passed into existing pipelines.

Almost all uses will rely on the SiffReader struct rather than the methods of the corrosiff module itself. The SiffReader struct can be constructed with open_siff and then used to read frames data, metadata, and other information. Some convenience methods are provided in the corrosiff module for converting .siff files to .tiff files or for reading frames directly for a one-time call if you don’t expect multiple interactions with the file.

Re-exports

pub use siffreader::SiffReader;

pub use siffreader::RegistrationDict;

pub use metadata::FrameMetadata;

Modules

metadata

The Metadata struct parses the microscope acquisition parameters and metadata from the various strings they are stored in scattered throughout the file. Most of the data is read from the nvfd – non-varying-frame-data – string, but information about frame times comes from data just before each individual IFD (for example).

siffreader

The primary SiffReader object, which parses files and extracts interesting information and/or data.

Macros

unwrap_tag_as

The ValueType field can either be of size u32 or u64, depending on the file format. The unwrap_tag_as! macro is a convenience macro for unwrapping a tag and converting it to a specific type. Use: unwrap_tag_as!(ifd, tag, into)

Enums

ClockBase

CorrosiffError

The CorrosiffError class reflects the major types of errors that occur in this library.

FramesError

Errors that can occur due to frame processing problems, either from the file reader (the IOError variant) or the values of the requested frames (e.g. incompatible dimensions or out of bounds).

TiffMode

Enum for specifying the conversion mode of a .siff file to a .tiff file.

Functions

close_siff_extern

close_siff_extern(reader) frees the memory allocated for the SiffReader object in lieu of freeing the memory in C.

get_frames

get_frames(path, frames, registration) returns the intensity data of the specified frames from the .siff file, with optional in-place registration.

get_frames_extern

NOT IMPLEMENTED

open_siff

open_siff(filename) opens a .siff file or a ScanImage-Flim .tiff file, reads the data, and returns a SiffReader object.

open_siff_extern

open_siff_extern(filename, len) opens a .siff file or a ScanImage-Flim .tiff file and returns a functional SiffReader object pointer – with responsibility for memory management on the part of the caller.

siff_to_tiff

siff_to_tiff(filename, mode) converts a .siff file to a .tiff file, using the specified mode. If the mode is TiffMode::ScanImage, the main metadata remains unconverted, and it uses the standard ScanImage format. If the mode is TiffMode::OME, the metadata is converted to the OME-TIFF format.

test_extern

time_axis_epoch

time_axis epoch

time_axis_epoch_both

time_axis_epoch_both

time_axis_experiment

time_axis_experiment