Function corrosiff::get_frames

pub fn get_frames(
    file_path: &str,
    frames: &[u64],
    registration: Option<&HashMap<u64, (i32, i32)>>
) -> Result<Array3<u16>, CorrosiffError>

get_frames(path, frames, registration) returns the intensity data of the specified frames from the .siff file, with optional in-place registration.

Arguments

file_path - A string slice that holds the name of the file to open

frames - A slice of u64 values specifying the frames to read

registration - An optional HashMap<u64, (i32, i32)> specifying the pixel shifts for each frame. If not specified, the frames are read unregistered.

Returns

Result<Array3<u16>, CorrosiffError> - An Array3<u16> containing the intensity data of the specified frames of size (frames, height, width).

Example

use corrosiff::{open_siff, get_frames};
    
let frames = vec![0, 1, 2, 3];
// No pixel shifts
let frames_array = get_frames("file.siff", &frames, None).unwrap();
 
// 256 lines, 128 pixels per line, 4 frames.
assert_eq!(frames_array.shape(), (4, 256, 128));

Errors

CorrosiffError::IOError - If there is an error reading the file or performing any conversion from the binary data, it is returned as an IOError

CorrosiffError::FramesError - If there is an error with the frames requested, e.g. inconsistent dimension or an out-of-bounds request, it is returned as a FramesError subtype.