Function corrosiff::siff_to_tiff

pub fn siff_to_tiff(
    filename: &str,
    _mode: TiffMode,
    save_path: Option<&String>
) -> Result<()>

siff_to_tiff(filename, mode) converts a .siff file to a .tiff file, using the specified mode. If the mode is TiffMode::ScanImage, the main metadata remains unconverted, and it uses the standard ScanImage format. If the mode is TiffMode::OME, the metadata is converted to the OME-TIFF format.

Arguments

filename - A string slice that holds the name of the file to open

mode - A string slice specifying the TiffMode enum to be used

save_path - An optional string slice that holds the path to save the converted file. If not specified, the file is saved in the same directory as the original file, with the same name but the extension .tiff.

Example

use corrosiff::{siff_to_tiff, TiffMode};
// Produces "file.tiff" in OME-TIFF format (not-yet implemented)
siff_to_tiff("file.siff", TiffMode::from_string_slice("OME"), None);
// Produces "file2.tiff" in ScanImage format
siff_to_tiff("file.siff", TiffMode::from_string_slice("ScanImage"), Some("file2.tiff"));