Function corrosiff::metadata::get_appended_text

pub fn get_appended_text<I: IFD, ReaderT: Read + Seek>(
    ifds: &[&I],
    reader: &mut ReaderT
) -> Vec<(u64, String, Option<f64>)>

Returns the appended text for each frame in a set of frames. This is text that is appended to the metadata string for each frame, and can be used to store additional information about the frame. This is often used to store information about the frame that is not easily stored in the metadata tags – timestamped events, notes, etc.

Arguments

ifds - A slice of IFD objects corresponding to the frames to get timestamps from

reader - An object with read access to the file

Returns

A Vec<(u64, String, Option<f64>)> containing the frame number (in terms of the IFDs passed in – NOT the actual frame number! it is the responsibility of the USER to correspond these values to the actual frame number), the appended text, and the timestamp (if one is present) for each frame in the order of the IFDs passed. The only entries in the vector are frames that actually have appended text.