Struct corrosiff::metadata::FrameMetadata

pub struct FrameMetadata {

    pub width: u64,
    pub height: u64,
    pub bits_per_sample: u16,
    pub compression: u16,
    pub photometric_interpretation: u16,
    pub end_of_ifd: u64,
    pub data_offset: u64,
    pub orientation: u16,
    pub samples_per_pixel: u16,
    pub rows_per_strip: u64,
    pub strip_byte_counts: u64,
    pub x_resolution: u64,
    pub y_resolution: u64,
    pub resolution_unit: u16,
    pub nvfd_address: u64,
    pub roi_address: u64,
    pub sample_format: u16,
    pub siff_compress: u16,
    pub metadata_string: String,

}

The Metadata struct is a Rust object that holds important or relevant metadata in a human-interpretable format.

It reads the tiff IFD data and also extracts a string containing more frame-specific metadata (e.g. frame timestamps). This struct is mostly for debugging – it’s a relatively inefficient way of extracting this information.

Fields
width: u64height: u64bits_per_sample: u16compression: u16photometric_interpretation: u16end_of_ifd: u64data_offset: u64orientation: u16samples_per_pixel: u16rows_per_strip: u64strip_byte_counts: u64x_resolution: u64y_resolution: u64resolution_unit: u16nvfd_address: u64roi_address: u64sample_format: u16siff_compress: u16metadata_string: String
Implementations

impl FrameMetadata

pub fn from_ifd<I: IFD>(ifd: &I) -> Self

Creates a new FrameMetadata object from an IFD. Because the IFD alone does not contain the long metadata string, this is not a complete metadata object. But its usable enough for most purposes.

Arguments

ifd - The IFD to create the metadata from

Returns

A new FrameMetadata object

Example

use crate::tiff::FileFormat;
use crate::metadata::FrameMetadata;
// `f` is a `File` object
let file_format = FileFormat::parse_filetype(&mut f).unwrap();
/* - snip - */
let ifds = file_format.get_ifd_vec(&mut f);
/* - snip - */
let metadata = FrameMetadata::from_ifd(&ifds[15]);

pub fn from_ifd_and_file<I: IFD, ReaderT: Read + Seek>( ifd: &I, reader: &mut ReaderT ) -> Result<Self, CorrosiffError>

Fully-parsed metadata, needs a file to read the metadata string from. Returns the reader to its original position if successful.

Arguments

ifd - The IFD to create the metadata from

reader - The reader to read the metadata string from

Returns

A new FrameMetadata object with a populated metadata string

Errors

CorrosiffError::IOError - If there is an error reading the metadata string due to an invalid Siff tag.

Example

use std::fs::File;
use std::io::BufReader;
use crate::tiff::FileFormat;
 
let f = File::open("file.siff").unwrap();
let reader = BufReader::new(f);
let file_format = FileFormat::parse_filetype(&mut reader).unwrap();
let ifds = file_format.get_ifd_vec(&mut reader);
let metadata = FrameMetadata::from_ifd_and_file(&ifds[15], &reader).unwrap();
 
println!("{}", metadata.metadata_string);

pub fn metadata_string<I: IFD, ReaderT: Read + Seek>( ifd: &I, reader: &mut ReaderT ) -> String

Reads a metadata string from the file without needing to create a FrameMetadata object. Returns the reader to its original position if successful.

Arguments

ifd - The IFD to read the metadata string from

reader - The reader to read the metadata string from

Returns

A String containing the metadata string

Example

use std::fs::File;
use std::io::BufReader;
use crate::tiff::FileFormat;
 
let f = File::open("file.siff").unwrap();
let reader = BufReader::new(f);
let file_format = FileFormat::parse_filetype(&mut reader).unwrap();
let ifds = file_format.get_ifd_vec(&mut reader);
let metadata_string = FrameMetadata::metadata_string(&ifds[15], &reader);
println!("{}", metadata_string);

pub fn frame_number_from_metadata_str(string: &str) -> u64

Public function for extracting the frame number from the metadata string without constructing an entire FrameMetadata object.

pub fn frame_time_experiment_from_metadata_str(string: &str) -> f64

Public function for extracting the frame time from the metadata string without constructing an entire FrameMetadata object.

pub fn frame_time_epoch_from_metadata_str(string: &str) -> u64

Public function for extracting the frame epoch from the metadata string without constructing an entire FrameMetadata object.

pub fn most_recent_system_time_from_metadata_str(string: &str) -> Option<u64>

Public function for extracting the most recent system time from the metadata string without constructing an entire FrameMetadata object.

pub fn sync_stamps_from_metadata_str(string: &str) -> u64

Public function for extracting the sync stamps from the metadata string without constructing an entire FrameMetadata object.

pub fn appended_text_from_metadata_str( string: &str ) -> Option<(String, Option<f64>)>

Public function for extracting the appended text from the metadata string without constructing an entire FrameMetadata object.

Returns

A tuple containing the full appended text, and the timestamp (if one is present). Is None for any string in which there is no appended text.