Struct corrosiff::siffreader::SiffReader

pub struct SiffReader { /* private fields */ }

A struct for reading a .siff file or a ScanImage-Flim .tiff file. Has methods which return arrays of image or FLIM

Implementations

impl SiffReader

pub fn open<P: AsRef<Path>>(filename: P) -> IOResult<Self>

Opens a file and returns a SiffReader object for interacting with the data if successful.

Arguments

filename - A string slice that holds the name of the file to open

Example

let reader = SiffReader::open("file.siff");

Errors

std::io::Error - If there is an error opening the file, it will be returned directly.

pub fn num_frames(&self) -> usize

Returns number of frames in the file (including flyback etc).

pub fn frames_vec(&self) -> Vec<u64>

pub fn image_dims(&self) -> Option<Dimensions>

pub fn filename(&self) -> String

Get value of the filename

Example

let reader = SiffReader::open("file.siff");
println!("{}", reader.filename());

pub fn nvfd(&self) -> String

Return the non-varying frame data, which contains all of the instrument settings

pub fn roi_string(&self) -> String

Return the mROI information as a string

pub fn is_bigtiff(&self) -> bool

Returns whether the file uses the BigTIFF format or the standard 32-bit TIFF format.

pub fn is_siff(&self) -> bool

Returns whether the file being read is a .siff (and contains lifetime information) or just a regular tiff file.

For now, this is implemented is a very unsophisticated manner – it simply checks whether the file ends in .siff!

TODO: Better!

pub fn get_frame_metadata( &self, frames: &[u64] ) -> Result<Vec<FrameMetadata>, CorrosiffError>

Return the metadata objects corresponding to each of the requested frames.

pub fn get_experiment_timestamps( &self, frames: &[u64] ) -> Result<Array1<f64>, CorrosiffError>

Return an array of timestamps corresponding to the experiment time of each frame requested (seconds since the onset of the acquisition).

Arguments

frames - A slice of u64 values corresponding to the frame numbers to retrieve

Returns

Result<Array1<f64>, CorrosiffError> - An array of f64 values corresponding to the timestamps of the frames requested (in units of seconds since beginning of the acquisition).

Example

let reader = SiffReader::open("file.siff");
let timestamps = reader.get_experiment_timestamps(&[0, 1, 2]);

pub fn get_epoch_timestamps_laser( &self, frames: &[u64] ) -> Result<Array1<u64>, CorrosiffError>

Return an array of timestamps corresponding to the epoch time of each frame requested computed using the number of laser pulses into the acquisition at the time of the frame trigger (and the estimated pulse rate).

This measurement has extremely low jitter but a fixed rate of drift from “true” epoch (as the system clock might read it).

Arguments

frames - A slice of u64 values corresponding to the frame numbers to retrieve

Returns

Result<Array1<u64>, CorrosiffError> - An array of u64 values corresponding to the epoch time of each frame

Example

let reader = SiffReader::open("file.siff");
let timestamps = reader.get_epoch_timestamps(&[0, 1, 2]);

pub fn get_epoch_timestamps_system( &self, frames: &[u64] ) -> Result<Array1<u64>, CorrosiffError>

Return an array of timestamps corresponding to the most recent epoch timestamped system call at the time of the frame trigger. This is the most accurate measure of system time because it does not drift, but the system is only queried about once a second, so there is high apparent jitter, with many consecutive frames sharing a value.

Arguments

frames - A slice of u64 values corresponding to the frame numbers to retrieve

Returns

Result<Array1<u64>, CorrosiffError> - An array of u64 values corresponding to the system time of each frame

Example

let reader = SiffReader::open("file.siff");
reader.get_epoch_timestamps_system(&[0, 1, 2]);

Errors

CorrosiffError::NoSystemTimestamps - If the system timestamps are not present in the file, this error is returned.

pub fn get_epoch_timestamps_both( &self, frames: &[u64] ) -> Result<Array2<u64>, CorrosiffError>

Return an array of timestamps corresponding to the two measurements of epoch time in the data: the laser clock synched one (low jitter, some drift) and the system call one (high jitter, no drift).

The two can be combined to allow much more reliable estimation of the timestamp of every frame trigger in absolute epoch time determined by the PTP system. These data are in nanoseconds since epoch.

Arguments

frames - A slice of u64 values corresponding to the frame numbers to retrieve

Returns

Result<Array2<u64>, CorrosiffError> - A 2D array of u64 values corresponding to the two epoch timestamps of each frame. The first row is laser_clock values, the second row is system_clock values. The system_clock changes only once a second or so, but this is much faster than the drift of the laser clock.

Example

let reader = SiffReader::open("file.siff");
reader.get_epoch_timestamps_both(&[0, 1, 2]);

Errors

CorrosiffError::NoSystemTimestamps - If the system timestamps are not present in the file, this error is returned.

CorrosiffError::DimensionsError(DimensionsError) - If the frames requested are out of bounds (the underlying DimensionsError is attached to this error)

pub fn get_appended_text( &self, frames: &[u64] ) -> Vec<(u64, String, Option<f64>)>

Returns a vector of all frames containing appended text. Each element of the vector is a tuple containing the frame number, the text itself, and the timestamp of the frame (if present).

Returns

(frame_number, text, Option<timestamp>)

pub fn get_frames_intensity( &self, frames: &[u64], registration: Option<&RegistrationDict> ) -> Result<Array3<u16>, CorrosiffError>

Return an array corresponding to the intensity data of the frames requested. The returned array is a 3D array, with the first dimension corresponding to the frame number, and the second and third dimensions corresponding to the y and x dimensions of the frame.

Arguments

frames - A slice of u64 values corresponding to the frame numbers to retrieve

registration - An optional HashMap<u64, (i32, i32)> which contains the pixel shifts for each frame. If this is None, the frames are read unregistered (runs faster).

Example

let reader = SiffReader::open("file.siff");
let frames = reader.get_frames_intensity(
    &[0, 1, 2],
    None
);

Errors

FramesError::DimensionsError(DimensionsError) - If the frames requested are out of bounds or do not all share the same shape (the underlying DimensionsError is attached to this error)

FramesError::IOError(_) - If there is an error reading the file (with the underlying error attached)

FramesError::RegistrationFramesMissing - If registration is used, and the registration values are missing for some frames

Returns

Result<Array3<u16>, FramesError> - A 3D array of u16 values corresponding to the intensity data of the frames requested.

pub fn get_frames_flim( &self, frames: &[u64], registration: Option<&RegistrationDict> ) -> Result<(Array3<f64>, Array3<u16>), CorrosiffError>

Return two arrays: the intensity (photon counts) and the empirical lifetime (in arrival time bins) of each pixel of the frames requested. corresponding to the y and x dimensions of the frame. The lifetime array is a 3D f64 array with the first dimension corresponding to the frame number, and the second and third dimensions corresponding to the y and x dimensions of the frame. The intensity array is a 3D u16 array with the same shape. If registration is None, the frames are read unregistered, otherwise they are registered in place.

Arguments

frames - A slice of u64 values corresponding to the frame numbers to retrieve

registration - An optional HashMap<u64, (i32, i32)> which contains the pixel shifts for each frame. If this is None, the frames are read unregistered (runs faster).

Returns

Result<(Array3<f64>, Array3<u16>), CorrosiffError> - A tuple containing the lifetime and intensity arrays of the frames requested (in that order). The lifetime array is the empirical lifetime in units of arrival time bins of the MultiHarp – meaning for every frame, each pixel contains the average arrival time of all photons in that pixel during that frame.

Example

let reader = SiffReader::open("file.siff");
let (lifetime, intensity) = reader.get_frames_flim(
   &[0, 1, 2],
  None
);
let intensity_alone = reader.get_frames(
  &[0, 1, 2],
 None
);
assert_eq!(intensity, intensity_alone);
// intensity is a 3D array of u16 values
// lifetime is a 3D array of f64 values in units of
// arrival time bins of the MultiHarp.

Errors

CorrosiffError::DimensionsError(DimensionsError) - If the frames requested are out of bounds or do not all share the same shape (the underlying DimensionsError is attached to this error)

CorrosiffError::FramesError(FramesError::IOError(_)) - If there is an error reading the file (with the underlying error attached)

CorrosiffError::FramesError(FramesError::RegistrationFramesMissing) - If registration is used, and the registration values are missing for some frames

See also

get_frames_intensity - for just the intensity data

get_histogram to pool all photons for a frame into a histogram

Panics

???

pub fn get_photon_stream( &self, frames: &[u64], registration: Option<&RegistrationDict> ) -> Result<Array1<u64>, CorrosiffError>

Returns a 1D array of u64 values corresponding to the photon stream of the frames requested, rather than converting the data into an array of intensity (or arrival) values.

Arguments

frames - A slice of u64 values corresponding to the frame numbers to retrieve

registration - An optional HashMap<u64, (i32, i32)> which

Panics

It’s not implemented yet!

pub fn get_histogram( &self, frames: &[u64] ) -> Result<Array2<u64>, CorrosiffError>

Returns a 2D array of u64 values corresponding to the number of photons in each bin of the arrival time histogram summing across ALL photons in the frames requested (not masked!).

Arguments

frames - A slice of u64 values corresponding to the frame numbers to retrieve

Returns

histogram - An ndarray::Array2<u64> with the first dimension equal to the number of frames and the second dimension equal to the number of arrival time bins of the histogram (read from the .siff metadata).

Example

let reader = SiffReader::open("file.siff");
let hist = reader.get_histogram(&[0, 1, 2]);
 
assert_eq!(reader.metadata().picoseconds_per_bin(), 20)
 
// 629 bins * 20 picoseconds per bin = 12.58 nanoseconds, ~80 MHz
assert_eq!(hist.shape(), &[3, 629]);

pub fn get_histogram_mask( &self, frames: &[u64], mask: &ArrayView2<'_, bool>, registration: Option<&RegistrationDict> ) -> Result<Array2<u64>, CorrosiffError>

Extracts a framewise photon arrival time histogram restricted only to the pixels in the region of interest.

Returns an array of u64 values corresponding to the number of photons in each bin of the arrival time histogram summing across all photons within the ROI for each frame.

Arguments

frames - The frames to apply the mask to

mask - A 2D boolean array with the same shape as the frames’ y and x dimensions. The ROI is true for the pixels that should be pooled to produce the histogram

registration - An optional HashMap<u64, (i32, i32)> which contains the pixel shifts for each frame. If this is None, the frames are read unregistered (runs faster).

Returns

Result<Array2<u64>, CorrosiffError> - An array of u64 values corresponding to the number of photons in each bin of the arrival time histogram for each frame requested. Shape is (frames.len(), num_bins).

pub fn sum_roi_flat( &self, roi: &ArrayView2<'_, bool>, frames: &[u64], registration: Option<&RegistrationDict> ) -> Result<Array1<u64>, CorrosiffError>

Sums the intensity of the frames requested within the region of interest (ROI) specified by the boolean array roi. The ROI should be the same shape as the frames’ y and x dimensions

Arguments

roi - A 2D boolean array with the same shape as the frames’ y and x dimensions. The ROI is a mask which will be used to sum the intensity of the frames requested.

frames - A slice of u64 values corresponding to the frame numbers to retrieve

registration - An optional HashMap<u64, (i32, i32)> which contains the pixel shifts for each frame. If this is None, the frames are read unregistered (runs faster).

Returns

Result<Array1<u64>, CorrosiffError> - An array of u64 values corresponding to the sum of the intensity of the frames requested within the ROI specified.

Example

let reader = SiffReader::open("file.siff");
let roi = Array2::<bool>::from_elem((512, 512), true);
// Set the ROI to false in the middle
roi.slice(s![200..300, 200..300]).fill(false);
 
let sum = reader.sum_roi_flat(&roi, &[0, 1, 2], None);

Errors

CorrosiffError::DimensionsError(DimensionsError) - If the frames requested are out of bounds, do not share the same shape, or the ROI does not share the same shape as the frames (the underlying DimensionsError is attached to this error)

See also

sum_roi_volume - for a 3D ROI mask

sum_rois_flat - for a set of 2D ROI masks

sum_roi_volume - for a set of 3D ROI masks

pub fn sum_roi_volume( &self, roi: &ArrayView3<'_, bool>, frames: &[u64], registration: Option<&RegistrationDict> ) -> Result<Array1<u64>, CorrosiffError>

Sums the intensity of the frames requested within the region of interest (ROI) specified by the boolean array roi. The ROI should have final two dimensions the same as the frames’ y and x dimensions, and the first dimension will be looped over while cycling through frames (i.e. frame 1 will be masked by the first plane of the ROI, frame 2 by the second plane, etc).

Arguments

roi - A 3D boolean array with the first dimension equal to the number of planes in the ROI, and the second and third dimensions corresponding to the y and x dimensions of the frames. The ROI is a mask which will be used to sum the intensity of the frames requested. Each plane is cycled through in parallel to the frames.

frames - A slice of u64 values corresponding to the frame numbers to retrieve

registration - An optional HashMap<u64, (i32, i32)> which contains the pixel shifts for each frame. If this is None, the frames are read unregistered (runs faster).

Returns

Result<Array1<u64>, CorrosiffError> - An array of u64 values corresponding to the sum of the intensity of the frames requested within the ROIs specified.

Example

let reader = SiffReader::open("file.siff");
let roi = Array3::<bool>::from_elem((3, 512, 512), true);
// Set the ROI to false in the middle
roi.slice(s![.., 200..300, 200..300]).fill(false);
 
let sum = reader.sum_roi_volume(&roi, &[0, 1, 2], None);
 
// The sum will be a 1D array of u64 values corresponding to the
// sum of the intensity of the frames requested within the ROIs specified.

Errors

CorrosiffError::DimensionsError(DimensionsError) - If the frames requested are out of bounds, do not share the same shape, or the ROI does not share the same shape as the frames (the underlying DimensionsError is attached to this error)

See also

sum_roi_flat - for a 2D ROI mask

sum_rois_flat - for a set of 2D ROI masks

sum_rois_volume - for a set of 3D ROI masks

pub fn sum_rois_flat( &self, rois: &ArrayView3<'_, bool>, frames: &[u64], registration: Option<&RegistrationDict> ) -> Result<Array2<u64>, CorrosiffError>

Sums a collection of masks over each frame requested and returns the sum of the intensity of the frames requested within each mask. Each ROI should have the same shape as the frames’ y and x dimensions. Runs slightly slower than applying the mask to just one ROI, so the gains scale as ~n_masks.

Arguments

rois - A 3D boolean array with the first dimension equal to the number of ROIs, and the second and third dimensions corresponding to the y and x dimensions of the frames. The ROIs are masks which will be used to sum the intensity of the frames requested.

frames - A slice of u64 values corresponding to the frame numbers to retrieve

registration - An optional HashMap<u64, (i32, i32)> which contains the pixel shifts for each frame. If this is None, the frames are read unregistered (runs faster).

Returns

Result<Array2<u64>, CorrosiffError> - An array of u64 values corresponding to the sum of the intensity of the frames requested within each ROI specified. Shape is (frames.len(), rois.len())

Example

let reader = SiffReader::open("file.siff");
let rois = Array3::<bool>::from_elem((3, 512, 512), true);
// TODO FINISH

See also

sum_roi_flat - for a 2D ROI mask

sum_rois_volume - for a set of 3D ROI masks

sum_roi_volume - for a single 3D ROI mask

pub fn sum_rois_volume( &self, rois: &ArrayView4<'_, bool>, frames: &[u64], registration: Option<&RegistrationDict> ) -> Result<Array2<u64>, CorrosiffError>

Sums a collection of 3d masks over each frame requested and returns the sum of the intensity of the frames requested within each mask. Each ROI (the last 2 dimensions) should have the same shape as the frames’ y and x dimensions. Iterates across the dimension 1 (i.e. the second dimension of the rois array) alongside the frames, so for each mask (masks are indexed along the slowest dimension) the 1st frame is applied to the the first plane of the mask, the 2nd frame to the second plane, and the nth frame to the n % roi.dim().1 plane. Runs slightly slower than applying the mask to just one ROI, so the gains scale as ~n_masks.

Arguments

rois - A 4D boolean array with the first dimension equal to the number of ROIs, the second dimension corresponding to each z plane for each mask, and the third and fourth dimensions corresponding to the y and x dimensions of the frames. The ROIs are masks which will be used to sum the intensity of the frames requested. Each plane is cycled through in parallel to the frames.

frames - A slice of u64 values corresponding to the frame numbers to retrieve

registration - An optional HashMap<u64, (i32, i32)> which contains the pixel shifts for each frame. If this is None, the frames are read unregistered (runs faster).

Returns

Result<Array2<u64>, CorrosiffError> - An array of u64 values corresponding to the sum of the intensity of the frames requested within each ROI specified. Shape is (frames.len(), rois.len())

Example

 
/* -- snip -- */
// 6 masks, 3 planes, 512 y pixels, 256 x pixels
let masks = Array4::<bool>::from_elem((6, 3, 512, 256), true);
// Set the ROIs to random values
use rand
masks.mapv_inplace(|_| rand::random::<bool>());
 
let frames = vec![0,1,2,3,4,5,6];
 
let masks_summed_at_once = reader.sum_rois_volume(
   &masks,
   &frames,
   None
)
 
// TODO

See also

sum_roi_flat - for a 2D ROI mask

sum_rois_flat - for a set of 2D ROI masks

sum_roi_volume - for a single 3D ROI mask

pub fn sum_roi_flim_flat( &self, roi: &ArrayView2<'_, bool>, frames: &[u64], registration: Option<&RegistrationDict> ) -> Result<(Array1<f64>, Array1<u64>), CorrosiffError>

Computes the empirical lifetime of the region masked by the ROI specified by the boolean array roi. The ROI should have the same shape as the frames’ y and x dimensions. Returns the lifetime and intensity data in a tuple, since it’s rare that you can use the lifetime without needing to refer to the intensity as well

Arguments

roi - A 2D boolean array with the same shape as the frames’ y and x dimensions. The ROI is a mask determining which pixels of each frame should be used to compute the lifetime.

frames - A slice of u64 values corresponding to the frame numbers to retrieve

registration - An optional HashMap<u64, (i32, i32)> which contains the pixel shifts for each frame. If this is None, the frames are read unregistered (runs faster).

Returns

Result<(Array1<f64>, Array1<u64>), CorrosiffError> - A tuple containing the lifetime and intensity data for the frames requested (respectively), each of shape (frames.len(),). The lifetime is in units of arrival time bins, and to transform to picoseconds, multiply by the picoseconds_per_bin value in the metadata. The intensity data is photon counts in the ROI (as in sum_roi_flat).

Example

// TODO

Errors

CorrosiffError::DimensionsError(DimensionsError) - If the frames requested are out of bounds, do not share the same shape, or the ROI does not share the same shape as the frames (the underlying DimensionsError is attached to this error)

See also

sum_roi_flim_volume - for a 3D ROI mask

sum_rois_flim_flat - for a set of 2D ROI masks

sum_rois_flim_volume - for a set of 3D ROI masks

pub fn sum_roi_flim_volume( &self, roi: &ArrayView3<'_, bool>, frames: &[u64], registration: Option<&RegistrationDict> ) -> Result<(Array1<f64>, Array1<u64>), CorrosiffError>

Computes the empirical lifetime of the region masked by the 3d ROI specified by the boolean array roi. The ROIs slowest axis is the slice axis, and the last two dimensions should have the same shape as the frames’ y and x dimensions. The z planes will be iterated through alongside the frames, so that roi.slice(s![0, .., ..]) will be applied to the first frame, roi.slice(s![1, .., ..]) to the second frame, in general the n % roi.dim().0 plane to the nth frame.

Arguments

roi - A 3D boolean array with the first dimension equal to the number of planes in the ROI, and the second and third dimensions corresponding to the y and x dimensions of the frames. The ROI is a mask which will be used to sum the intensity of the frames requested and compute the mean arrival time of the photons in the ROI.

frames - A slice of u64 values corresponding to the frame numbers to retrieve

registration - An optional HashMap<u64, (i32, i32)> which contains the pixel shifts for each frame. If this is None, the frames are read unregistered (runs faster).

Returns

Result<(Array1<f64>, Array1<u64>), CorrosiffError> - A tuple containing the lifetime and intensity data for the frames requested (respectively), each of shape (frames.len(),). The lifetime is in units of arrival time bins, and to transform to picoseconds, multiply by the picoseconds_per_bin value in the metadata. The intensity data is photon counts in the ROI (as in sum_roi_flat).

Example

// TODO

Errors

CorrosiffError::DimensionsError(DimensionsError) - If the frames requested are out of bounds, do not share the same shape, or the ROI does not share the same shape as the frames (the underlying DimensionsError is attached to this error)

See also

sum_roi_flim_flat - for a 2D ROI mask

sum_rois_flim_flat - for a set of 2D ROI masks

sum_rois_flim_volume - for a set of 3D ROI masks

pub fn sum_rois_flim_flat( &self, rois: &ArrayView3<'_, bool>, frames: &[u64], registration: Option<&RegistrationDict> ) -> Result<(Array2<f64>, Array2<u64>), CorrosiffError>

Sums a collection of masks over each frame requested and returns the empirical lifetime and intensity of the frames requested within each mask. Each ROI should have the same shape as the frames’ y and x dimensions. Runs slightly slower than applying the mask to just one ROI, so the gains scale as ~n_masks.

Arguments

rois - A 3D boolean array with the first dimension equal to the number of ROIs, and the second and third dimensions corresponding to the y and x dimensions of the frames. The ROIs are masks which will be used to sum the intensity of the frames requested.

frames - A slice of u64 values corresponding to the frame numbers to retrieve

registration - An optional HashMap<u64, (i32, i32)> which contains the pixel shifts for each frame. If this is None, the frames are read unregistered (runs faster).

Returns

Result<(Array2<f64>, Array2<u64>), CorrosiffError> - A tuple containing the lifetime and intensity data for the frames requested within each ROI specified. The first element of the tuple is the lifetime data, and the second element is the intensity data. The shape of each array is (frames.len(), rois.len())

Example

let reader = SiffReader::open("file.siff");
let rois = Array3::<bool>::from_elem((3, 512, 512), true);
// TODO FINISH

See also

sum_roi_flim_flat - for a 2D ROI mask

sum_rois_flim_volume - for a set of 3D ROI masks

sum_roi_flim_volume - for a single 3D ROI mask

pub fn sum_rois_flim_volume( &self, rois: &ArrayView4<'_, bool>, frames: &[u64], registration: Option<&RegistrationDict> ) -> Result<(Array2<f64>, Array2<u64>), CorrosiffError>

Sums a collection of 3d masks over each frame requested and returns the empirical lifetime and total intensity of the frames requested within each mask. Each ROI (the last 2 dimensions) should have the same shape as the frames’ y and x dimensions. Iterates across the dimension 1 (i.e. the second dimension of the rois array) alongside the frames, so for each mask (masks are indexed along the slowest dimension) the 1st frame is applied to the the first plane of the mask, the 2nd frame to the second plane, and the nth frame to the n % roi.dim().1 plane. Runs slightly slower than applying the mask to just one ROI, so the gains scale as ~n_masks.

Arguments

rois - A 4D boolean array with the first dimension equal to the number of ROIs, the second dimension corresponding to each z plane for each mask, and the third and fourth dimensions corresponding to the y and x dimensions of the frames. The ROIs are masks which will be used to sum the intensity of the frames requested. Each plane is cycled through in parallel to the frames.

frames - A slice of u64 values corresponding to the frame numbers to retrieve

registration - An optional HashMap<u64, (i32, i32)> which contains the pixel shifts for each frame. If this is None, the frames are read unregistered (runs faster).

Returns

Result<(Array2<f64>, Array2<u64>), CorrosiffError> - A tuple of arrays, the first of which contains the lifetime data for the frames (in units of arrival time bins) and the second of which contains the intensity data for the frames requested within each ROI specified. The shape of each array is (frames.len(), rois.len()). To convert the lifetime data to picoseconds, multiply by the picoseconds_per_bin value in the metadata.

Example

 
/* -- snip -- */
// 6 masks, 3 planes, 512 y pixels, 256 x pixels
let masks = Array4::<bool>::from_elem((6, 3, 512, 256), true);
// Set the ROIs to random values
use rand
masks.mapv_inplace(|_| rand::random::<bool>());
 
let frames = vec![0,1,2,3,4,5,6];
 
let masks_summed_at_once = reader.sum_rois_volume(
   &masks,
   &frames,
   None
)
 
// TODO

See also

sum_roi_flim_flat - for a 2D ROI mask

sum_rois_flim_flat - for a set of 2D ROI masks

sum_roi_flim_volume - for a single 3D ROI mask

sum_rois_volume - the intensity-only version.